# src/app/authentication/docs/swagger.yaml
definitions:
...
  models.ForgotPasswordRequest:
    properties:
      NewPassword:
        example: sample11110
        maxLength: 20
        minLength: 8
        type: string
      email:
        example: testUser@gmail.com
        type: string
      panCardNumber:
        example: abgjhi6789
        type: string
    required:
    - email
    - panCardNumber
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: api for Stock Broker using gin and gorm
  title: Stock Broker Application
  version: "2.0"
paths:
...
  /v1/forgot-password:
    post:
      consumes:
      - application/json
      description: Updates user credentials based on the provided request.
      operationId: update-credentials
      parameters:
      - description: Forgot Password Request JSON
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password updated successfully
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update user credentials
      tags:
      - Forgot Password
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"